requirements.txt
flask
google-cloud-artifact-registry
gunicorn